Moving logging code into config.py
